Local changes

+++ courses-rest-services/src/main/java/com/fromero/coursesapi/service/CourseService.java
@@ -9,6 +9,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.util.Streamable;
 import org.springframework.stereotype.Service;
 
+import com.fromero.coursesapi.dto.CourseDTO;
+import com.fromero.coursesapi.error.ApiReturnMessage;
+import com.fromero.coursesapi.errorhandling.BusinessErrorException;
 import com.fromero.coursesapi.model.Course;
 import com.fromero.coursesapi.repository.CourseRepository;
 
@@ -30,14 +33,23 @@ public class CourseService implements Serializable {
     }
 
     public Course getCourse(Integer id) {
+    	if (id == null) {
+			throw new BusinessErrorException(ApiReturnMessage.ERR_ID_NOT_INFORMED);
+		}    	
         return courseRepository.findById(id).get();
     }
 
     public void deleteCourse(Integer id) {
+    	if (id == null) {
+			throw new BusinessErrorException(ApiReturnMessage.ERR_ID_NOT_INFORMED);
+		}    	
     	courseRepository.deleteById(id);
     }
 
-	public List<Course> listAllCoursesByStudent(Integer id) {
+	public List<CourseDTO> listAllCoursesByStudent(Integer id) {
+    	if (id == null) {
+			throw new BusinessErrorException(ApiReturnMessage.ERR_ID_NOT_INFORMED);
+		}    	
 		return courseRepository.listCoursesByStudent(id);
 	}
